refactor: move ApiRequestError to Error module
