+++ src/PostgREST/ApiRequest/Types.hs
@@ -5,7 +5,6 @@ module PostgREST.ApiRequest.Types
   , Cast
   , Depth
   , EmbedParam(..)
-  , ApiRequestError(..)
   , EmbedPath
   , Field
   , Filter(..)
@@ -25,9 +24,6 @@ module PostgREST.ApiRequest.Types
   , OrderDirection(..)
   , OrderNulls(..)
   , OrderTerm(..)
-  , QPError(..)
-  , RaiseError(..)
-  , RangeError(..)
   , SingleVal
   , IsVal(..)
   , SimpleOperator(..)
@@ -36,12 +32,7 @@ module PostgREST.ApiRequest.Types
   , SelectItem(..)
   ) where
 
-import PostgREST.MediaType                (MediaType (..))
-import PostgREST.SchemaCache.Identifiers  (FieldName,
-                                           QualifiedIdentifier)
-import PostgREST.SchemaCache.Relationship (Relationship,
-                                           RelationshipsMap)
-import PostgREST.SchemaCache.Routine      (Routine (..))
+import PostgREST.SchemaCache.Identifiers (FieldName)
 
 import Protolude
 
@@ -68,49 +59,6 @@ data SelectItem
     , selJoinType :: Maybe JoinType
     }
   deriving (Eq, Show)
-
-data ApiRequestError
-  = AggregatesNotAllowed
-  | AmbiguousRelBetween Text Text [Relationship]
-  | AmbiguousRpc [Routine]
-  | MediaTypeError [ByteString]
-  | InvalidBody ByteString
-  | InvalidFilters
-  | InvalidPreferences [ByteString]
-  | InvalidRange RangeError
-  | InvalidRpcMethod ByteString
-  | NotFound
-  | NoRelBetween Text Text (Maybe Text) Text RelationshipsMap
-  | NoRpc Text Text [Text] MediaType Bool [QualifiedIdentifier] [Routine]
-  | NotEmbedded Text
-  | PutLimitNotAllowedError
-  | QueryParamError QPError
-  | RelatedOrderNotToOne Text Text
-  | SpreadNotToOne Text Text
-  | UnacceptableFilter Text
-  | UnacceptableSchema [Text]
-  | UnsupportedMethod ByteString
-  | ColumnNotFound Text Text
-  | GucHeadersError
-  | GucStatusError
-  | PutMatchingPkError
-  | SingularityError Integer
-  | PGRSTParseError RaiseError
-  | MaxAffectedViolationError Integer
-  deriving Show
-
-data QPError = QPError Text Text
-  deriving Show
-data RaiseError
-  = MsgParseError ByteString
-  | DetParseError ByteString
-  | NoDetail
-  deriving Show
-data RangeError
-  = NegativeLimit
-  | LowerGTUpper
-  | OutOfBounds Text Text
-  deriving Show
 
 type NodeName = Text
 type Depth = Integer
